refactor(codegen): share AbiWordValidator with calldata decode
Move AbiWordValidator and enum_cleanup_mask from transform/lower_abi.rs
to mir/abi.rs next to AbiParamType, and make the calldata decoding
helpers in lower/function/abi_calls.rs use them instead of maintaining
a second copy of the masks, signextends, and bool canonicalization.
Bit-identical output; full uitest and codegen unit tests pass.
